refactor: move current home page to dashboard and add a landing page
