+++ src/routes/dashboard/(components)/GetStartedComponent.svelte
