+++ src/routes/dashboard/(components)/WorkoutProgressionChart.svelte
